no max drag speed lock but limit init frame drag
the old solution limited dragging speed to a constant rate which even with damping and smoothing would feel not great. plus unnecessary, so to really adress the issue the initial frame of dragging which goes to a higher than expected value on mobile pointer vs normal mouse cursor, just disregard the initial frame. sensitivity scaling is still applied relative to distance from object caluclated once at the start of the dragging.